+++ jasmine/throw-error.js
